Made useRaceSelection and useFilteredData hooks
